Move homepage code to main.js file
