+++ src/App.js
@@ -1,5 +1,6 @@
 import { Routes, Route } from "react-router-dom";
-import Home from "./pages/HomePage/HomePage";
+// import Home from "./pages/HomePage/HomePage";
+import Home from "./main.js"
 
 function App() {
   return (
